Avoid double copy of nvim and terminator

+++ dotfilesLink.sh
@@ -7,6 +7,12 @@ ln -si $HOME/dotfiles/xinitrc $HOME/.xinitrc
 ln -si $HOME/dotfiles/zshrc $HOME/.zshrc
 ln -si $HOME/dotfiles/xmonad.hs $HOME/.xmonad/xmonad.hs
 ln -si $HOME/dotfiles/xmobarrc $HOME/.xmobarrc
-ln -si $HOME/dotfiles/terminator/ $HOME/.config/terminator
 ln -si $HOME/dotfiles/stalonetrayrc $HOME/.stalonetrayrc
-ln -si $HOME/dotfiles/nvim/ $HOME/.config/nvim
+
+if [ ! -e $HOME/.config/nvim ]; then
+    ln -si $HOME/dotfiles/nvim/ $HOME/.config/nvim
+fi
+
+if [ ! -e $HOME/.config/terminator ]; then
+    ln -si $HOME/dotfiles/terminator/ $HOME/.config/terminator
+fi
